(feat: 댓글 컴포넌트 기능 수정 및 에러 픽스)
